+++ src/main/java/org/zalando/nakadi/exceptions/runtime/SubscriptionUpdateConflictException.java
@@ -1,0 +1,7 @@
+package org.zalando.nakadi.exceptions.runtime;
+
+public class SubscriptionUpdateConflictException extends MyNakadiRuntimeException1 {
+    public SubscriptionUpdateConflictException(final String message) {
+        super(message);
+    }
+}