+++ src/main/java/org/zalando/nakadi/service/subscription/SubscriptionValidationService.java
@@ -16,10 +16,12 @@ import org.zalando.nakadi.exceptions.runtime.InconsistentStateException;
 import org.zalando.nakadi.exceptions.runtime.NoEventTypeException;
 import org.zalando.nakadi.exceptions.runtime.RepositoryProblemException;
 import org.zalando.nakadi.exceptions.runtime.ServiceTemporarilyUnavailableException;
+import org.zalando.nakadi.exceptions.runtime.SubscriptionUpdateConflictException;
 import org.zalando.nakadi.exceptions.runtime.TooManyPartitionsException;
 import org.zalando.nakadi.exceptions.runtime.WrongInitialCursorsException;
 import org.zalando.nakadi.exceptions.runtime.WrongStreamParametersException;
 import org.zalando.nakadi.repository.EventTypeRepository;
+import org.zalando.nakadi.service.AuthorizationValidator;
 import org.zalando.nakadi.service.CursorConverter;
 import org.zalando.nakadi.service.timeline.TimelineService;
 import org.zalando.nakadi.view.SubscriptionCursorWithoutToken;
@@ -28,6 +30,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.function.Function;
 import java.util.stream.Collectors;
@@ -41,16 +44,19 @@ public class SubscriptionValidationService {
     private final TimelineService timelineService;
     private final int maxSubscriptionPartitions;
     private final CursorConverter cursorConverter;
+    private final AuthorizationValidator authorizationValidator;
 
     @Autowired
     public SubscriptionValidationService(final TimelineService timelineService,
                                          final EventTypeRepository eventTypeRepository,
                                          final NakadiSettings nakadiSettings,
-                                         final CursorConverter cursorConverter) {
+                                         final CursorConverter cursorConverter,
+                                         final AuthorizationValidator authorizationValidator) {
         this.timelineService = timelineService;
         this.eventTypeRepository = eventTypeRepository;
         this.maxSubscriptionPartitions = nakadiSettings.getMaxSubscriptionPartitions();
         this.cursorConverter = cursorConverter;
+        this.authorizationValidator = authorizationValidator;
     }
 
     public void validateSubscription(final SubscriptionBase subscription)
@@ -74,6 +80,28 @@ public class SubscriptionValidationService {
         if (subscription.getReadFrom() == SubscriptionBase.InitialPosition.CURSORS) {
             validateInitialCursors(subscription, allPartitions);
         }
+        // Verify that subscription authorization object is valid
+        authorizationValidator.validateAuthorization(subscription.getAuthorization());
+    }
+
+    public void validateSubscriptionChange(final Subscription old, final SubscriptionBase newValue)
+            throws SubscriptionUpdateConflictException {
+        if (!Objects.equals(newValue.getConsumerGroup(), old.getConsumerGroup())) {
+            throw new SubscriptionUpdateConflictException("Not allowed to change subscription consumer group");
+        }
+        if (!Objects.equals(newValue.getEventTypes(), old.getEventTypes())) {
+            throw new SubscriptionUpdateConflictException("Not allowed to change subscription event types");
+        }
+        if (!Objects.equals(newValue.getOwningApplication(), old.getOwningApplication())) {
+            throw new SubscriptionUpdateConflictException("Not allowed to change owning application");
+        }
+        if (!Objects.equals(newValue.getReadFrom(), old.getReadFrom())) {
+            throw new SubscriptionUpdateConflictException("Not allowed to change read from");
+        }
+        if (!Objects.equals(newValue.getInitialCursors(), old.getInitialCursors())) {
+            throw new SubscriptionUpdateConflictException("Not allowed to change initial cursors");
+        }
+        authorizationValidator.validateAuthorization(old.getAuthorization(), newValue.getAuthorization());
     }
 
     public void validatePartitionsToStream(final Subscription subscription, final List<EventTypePartition> partitions) {
@@ -169,5 +197,4 @@ public class SubscriptionValidationService {
                     StringUtils.join(missingEventTypes, "', '")));
         }
     }
-
 }